+++ admin/src/main/resources/io/buoyant/admin/js/main-linkerd.js
@@ -35,11 +35,13 @@ require([
   linkerdDtabPlayground
 ) {
   // poor man's routing
-  if (window.location.pathname.indexOf("delegator") === 1) {
-    adminPage.initialize(true);
-    new linkerdDtabPlayground();
+  if (window.location.pathname.endsWith("/delegator")) {
+    adminPage.initialize(true).done(function() {
+      new linkerdDtabPlayground();
+    });
   } else {
-    adminPage.initialize();
-    new dashboard();
+    adminPage.initialize().done(function() {
+      new dashboard();
+    });
   }
 });