+++ admin/src/main/resources/io/buoyant/admin/js/src/router_clients.js
@@ -13,10 +13,8 @@ define([
     var EXPAND_CLIENT_THRESHOLD = 6;
 
     function assignColorsToClients(colors, clients) {
-      var colorIdx = 0;
-
-      return _.reduce(clients, function(clientMapping, client) {
-        clientMapping[client.label] = colors[colorIdx++ % colors.length];
+      return _.reduce(clients, function(clientMapping, client, idx) {
+        clientMapping[client.label] = colors[idx % colors.length];
         return clientMapping;
       }, {});
     }
@@ -29,14 +27,15 @@ define([
     return function (metricsCollector, routers, $clientEl, routerName, colors) {
       var clientContainerTemplate = Handlebars.compile(routerClientContainerTemplate);
 
-      var clientToColor = assignColorsToClients(colors, routers.clients(routerName));
-      var combinedClientGraph = CombinedClientGraph(metricsCollector, routers, routerName, $clientEl.find(".router-graph"), clientToColor);
       var clients = routers.clients(routerName);
+      var colorList = colors;
+      var clientToColor = assignColorsToClients(colorList, clients);
+      var combinedClientGraph = CombinedClientGraph(metricsCollector, routers, routerName, $clientEl.find(".router-graph"), clientToColor);
 
       var expandClients = shouldExpandClients(clients.length);
 
       var routerClients = _.map(clients, function(client) {
-        return initializeClient(client, expandClients);
+        return initializeClient(client);
       });
 
       if (routerClients.length == 0) {
@@ -45,7 +44,7 @@ define([
 
       routers.onAddedClients(addClients);
 
-      function initializeClient(client, shouldExpand) {
+      function initializeClient(client) {
         $clientEl.show();
         var colorsForClient = clientToColor[client.label];
         var $container = $(clientContainerTemplate({
@@ -56,29 +55,29 @@ define([
         var $chart = $container.find(".chart-container");
         var $toggle = $container.find(".client-toggle");
 
-        return RouterClient(metricsCollector, routers, client, $metrics, routerName, $chart, colorsForClient, $toggle, shouldExpand);
+        return RouterClient(metricsCollector, routers, client, $metrics, routerName, $chart, colorsForClient, $toggle, expandClients);
       }
 
       function addClients(addedClients) {
-        // reassign colors
-        clientToColor = assignColorsToClients(colors, routers.clients(routerName));
-
-        // update existing client colors
-        combinedClientGraph.updateColors(clientToColor);
-
-        _.each(routerClients, function(routerClient) {
-          routerClient.updateColors(clientToColor);
+        // filter new clients
+        var filteredClients = _.filter(addedClients, function(client) {
+          return client.router === routerName;
         });
 
-        var expandClients = shouldExpandClients(addedClients.length + routerClients.length);
+        // combine with existing clients
+        var combinedClients = routerClients.concat(filteredClients);
 
-        // add new clients
-        _.chain(addedClients)
-          .filter(function(client) { return client.router === routerName })
-          .each(function(clientForRouter) {
-            routerClients.push(initializeClient(clientForRouter, expandClients));
-          })
-          .value();
+        // reassign colors
+        clientToColor = assignColorsToClients(colorList, combinedClients);
+
+        // pass new colors to combined request graph, add new clients to graph
+        combinedClientGraph.updateColors(clientToColor);
+        combinedClientGraph.addClients(filteredClients);
+
+        // add new clients to dom
+        _.each(filteredClients, function(clientForRouter) {
+          routerClients.push(initializeClient(clientForRouter));
+        });
       }
     }
   })();