+++ linkerd/admin/src/main/scala/io/buoyant/linkerd/admin/AdminHandler.scala
@@ -8,6 +8,10 @@ import io.buoyant.linkerd.Build
 
 object AdminHandler extends HtmlView {
 
+  /**
+   * NOTE: this navigation must be kept in sync with the nav bar in:
+   * admin/src/main/resources/io/buoyant/admin/js/src/logger-overrides.js
+   */
   val navBar =
     s"""<nav class="navbar navbar-inverse">
       <div class="navbar-container">
@@ -19,16 +23,16 @@ object AdminHandler extends HtmlView {
             <span class="icon-bar"></span>
           </button>
 
-          <a class="navbar-brand-img" href="/">
-            <img alt="Logo" src="/files/images/linkerd-horizontal-white-transbg-vectorized.svg">
+          <a class="navbar-brand-img" href=".">
+            <img alt="Logo" src="files/images/linkerd-horizontal-white-transbg-vectorized.svg">
           </a>
         </div>
         <div id="navbar" class="collapse navbar-collapse">
           <ul class="nav navbar-nav">
-            <li><a href="/delegator">dtab</a></li>
-            <li><a href="/requests">requests</a></li>
-            <li><a href="/admin/logging">logging</a></li>
-            <li><a href="/help">help</a></li>
+            <li><a href="delegator">dtab</a></li>
+            <li><a href="requests">requests</a></li>
+            <li><a href="admin/logging">logging</a></li>
+            <li><a href="help">help</a></li>
           </ul>
           <ul class="nav navbar-nav navbar-right">
             <li class="dropdown">
@@ -61,17 +65,16 @@ object AdminHandler extends HtmlView {
     navbar: String = navBar
   ): String = {
     val cssesHtml = csses.map { css =>
-      s"""<link type="text/css" href="/files/css/$css" rel="stylesheet"/>"""
+      s"""<link type="text/css" href="files/css/$css" rel="stylesheet"/>"""
     }.mkString("\n")
-
     s"""
       <!doctype html>
       <html>
         <head>
           <title>linkerd admin</title>
-          <link type="text/css" href="/files/css/lib/bootstrap.min.css" rel="stylesheet"/>
-          <link type="text/css" href="/files/css/dashboard.css" rel="stylesheet"/>
-          <link rel="shortcut icon" href="/files/images/favicon.png" />
+          <link type="text/css" href="files/css/lib/bootstrap.min.css" rel="stylesheet"/>
+          <link type="text/css" href="files/css/dashboard.css" rel="stylesheet"/>
+          <link rel="shortcut icon" href="files/images/favicon.png" />
           <link href='https://fonts.googleapis.com/css?family=Source+Sans+Pro:400,300,600' rel='stylesheet' type='text/css'>
           $cssesHtml
         </head>
@@ -83,7 +86,7 @@ object AdminHandler extends HtmlView {
           </div>
           $tailContent
 
-          <script data-main="/files/js/main-linkerd" src="/files/js/lib/require.js"></script>
+          <script data-main="files/js/main-linkerd" src="files/js/lib/require.js"></script>
 
         </body>
       </html>